implemented U-net Generator
